TEST: Notices API Mock Test 추가
